bp::file::size() needs to return a uintmax_t rather than a size_t
git-svn-id: svn+ssh://corpsvnproxy/yahoo/BrowserPlus/public_platform/trunk@347 3fcf768e-b076-0410-b9b8-e8c34e2d470b

+++ src/libs/bputils/src/Http/Client/HttpTransaction_Windows.cpp
@@ -200,10 +200,10 @@ private:
 
     // progress info
     IListener*      m_pListener;
-    size_t          m_sendTotalBytes;
-    size_t          m_bytesSent;
-    size_t          m_receiveTotalBytes;
-    size_t          m_bytesReceived;
+    uintmax_t       m_sendTotalBytes;
+    uintmax_t       m_bytesSent;
+    uintmax_t       m_receiveTotalBytes;
+    uintmax_t       m_bytesReceived;
     bool            m_zeroProgressSent;
     double          m_lastProgressSent;
